--- src/pages/message/message.json
@@ -1,8 +1,0 @@
-{
-  "navigationBarTitleText": "我的消息",
-  "navigationBarBackgroundColor": "#3A82E6",
-  "navigationBarTextStyle": "white",
-  "usingComponents": {
-    "Loading": "../../components/Loading/Loading"
-  }
-}
