--- src/pages/message/message.scss
@@ -1,8 +1,0 @@
-@import "../../scss/variables";
-
-.message-card-avatar {
-  width: 94rpx;
-  height: 94rpx;
-  border-radius: 50%;
-  background-color: $bg-color;
-}
