--- src/pages/message/message.wxml
@@ -1,20 +1,0 @@
-<view class="padding-t-md">
-  <view class="margin-l-xl margin-r-xl">
-    <Loading wx:if="{{!messageList}}" />
-    <view wx wx:for="{{messageList}}" class="card">
-      <view class="card-block" data-id="{{item.messageId}}" bindtap="handleClickMessage">
-        <view class="flex-align-center padding-t-lg padding-b-lg">
-          <image class="message-card-avatar" mode="aspectFill" src="{{'../../images/avatar_default.jpeg'}}" />
-          <view class="flex-item padding-l-md">
-            <view class="padding-b-xs">
-              <text class="flex-item h1 text-title">{{item.fromUserName}}</text>
-            </view>
-            <view class="flex-align-center h5 activity-info">
-              <text class="text-title padding-l-xs padding-r-xs text-muted">{{item.payload}}</text>
-            </view>
-          </view>
-        </view>
-      </view>
-    </view>
-  </view>
-</view>
