+++ src/pages/notice/notice.json
@@ -1,7 +1,9 @@
 {
-  "navigationBarTitleText": "消息通知",
-  "enablePullDownRefresh": true,
+  "navigationBarTitleText": "我的消息",
+  "navigationBarBackgroundColor": "#3A82E6",
+  "navigationBarTextStyle": "white",
   "usingComponents": {
+    "Loading": "../../components/Loading/Loading",
     "NoticeItem": "../../components/NoticeItem/NoticeItem"
   }
 }
