+++ src/pages/notice/notice.wxml
@@ -1,13 +1,6 @@
-<view class="container padding-t-md">
-  <view class="card">
-    <view class="card-block">
-      <NoticeItem
-        wx:for="{{noticeList}}"
-        wx:key="messageId"
-        data="{{item}}"
-        data-id="{{item.fromUser}}"
-        bindtap="handleClickCard"
-      />
-    </view>
+<view class="padding-t-md">
+  <view class="margin-l-xl margin-r-xl">
+    <Loading wx:if="{{!messageList}}" />
+    <NoticeItem wx:key="messageId" wx:for="{{messageList}}" data="{{item}}" bindtap="handleClickMessage"/>
   </view>
 </view>
